lib: correct amqp_time_from_now behavior at {0, 0}
Change behavior of amqp_time_from_now to set to the current timestamp
when a struct timeval of {0, 0} is passed in instead of 0. This better
aligns with how one would expect the function to work.

As a byproduct this corrects an issue where amqp_consume_message would
not return AMQP_STATUS_HEARTBEAT_TIMEOUT when a tv of {0, 0} and the
heartbeat interval had passed.

Fixes #557 in a more concise way.